Span a non-indexed part-select's indices when its first bound is x, `a[1'bx : -2]` having read the second index as a width and asked MakeAllX for 4294967294 bits where §11.5.1 yields x
§11.5.1 gives the second expression of a part-select two meanings and spells the forms apart: in `vect[msb_expr : lsb_expr]` it is an index, and in `down_vect[lsb_base_expr +: width_expr]` it is a width, one that "shall be a positive constant integer expression". EvalUnknownIndexSelect's part-select arm consulted neither is_part_select_plus nor is_part_select_minus and took it as the width for both, so the span `a[1'bx : 4]` names was read off the number 4 and `a[1'bx : 6]` came back six bits wide where its indices name two.

A negative second index made that a memory-safety defect rather than a wrong width. `-2` is a unary minus over an unsized decimal, so §11.6.1 gives it 32 bits and the bound arrived as 4294967294; MakeLogic4Vec's `(width + 63) / 64` wrapped that in 32 bits to a word count of 0, and the vector returned claimed 4294967294 bits while holding no words at all. ResizeToWidth copied none of them, so `r = a[1'bx : -2]` on a `logic [3:0] r` read 4'b0000 -- the one answer §11.5.1 forbids, since 0 is what a 2-state out-of-range select yields and r is logic -- and Logic4Vec::ToString, which walks from width - 1 with no test of nwords, would have read a gigabyte past a zero-byte allocation had anything printed it. A second index between 65 and 2^31 allocated and filled in proportion to itself instead, 25 MB for `a[1'bx : 100000000]`, never freed before the arena is.

The arm now asks which form it holds. For the two indexed forms it reads the width through SelectBoundValue, which answers a negative one with the minimum of 1 rather than a large positive number. For the non-indexed form the first index is x and the span is not a number the expression states, so what bounds it is the declaration: §11.5.1 makes the first index the more significant end of the pair, and the widest such select the object admits runs from the known second index -- clamped into the base's range, since a bound outside it addresses no bit -- up to the most significant index there is. A second index at or past that end leaves the minimum of one bit, which is what the same clause gives a select "completely out of the address bounds".

MakeLogic4Vec rounds in 64 bits now on its own account. The wrap is silent for every width above 0xFFFFFFC0 and builds a vector whose width and nwords disagree, and this arm was the caller that reached it rather than the only one that could; every width a declaration can state rounds identically either way.

Closes #3533

+++ src/common/types.cpp
@@ -114,7 +114,14 @@ void Logic4Snapshot::Capture(const Logic4Vec& src) {
 }
 
 Logic4Vec MakeLogic4Vec(Arena& arena, uint32_t width) {
-  uint32_t nwords = (width + 63) / 64;
+  // The rounding is done in 64 bits because it overflows in 32: a width above
+  // 0xFFFFFFC0 wrapped `width + 63` back to a small number and answered a word
+  // count far below the one the width needs -- 0 for the widths just under the
+  // maximum -- leaving a vector whose width and nwords disagree and which
+  // Logic4Vec::ToString then reads out of its own allocation. Every width a
+  // declaration can state rounds identically either way.
+  uint32_t nwords =
+      static_cast<uint32_t>((static_cast<uint64_t>(width) + 63) / 64);
   auto* words = arena.AllocArray<Logic4Word>(nwords);
   return {width, nwords, words};
 }
